requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.11
django-filter==23.5
django-import-export==3.2.0
django-redis==5.4.0
django-redis-cache==3.0.1
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
MarkupPy==1.14
MarkupSafe==2.1.5
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.1.2
packaging==23.2
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
redis==3.5.3
requests==2.31.0
simplejson==3.19.2
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.2.1
xlrd==2.0.1
xlwt==1.3.0
